Telnet system refactoring: Moved many constructors to simpler newer patterns, refactored many Telnet command bytes to a helper class for readability instead of hard-coded values (and combined the DO/DONT/etc with those), optimized some static response patterns, and a bit of other minor cleanup.

+++ src/Server/Interfaces/ITelnetOption.cs
@@ -14,7 +14,7 @@ namespace WheelMUD.Server.Interfaces
         string Name { get; }
 
         /// <summary>Gets the code number of the option.</summary>
-        int OptionCode { get; }
+        byte OptionCode { get; }
 
         /// <summary>Attempt to enable the option.</summary>
         void Enable();
